app/pyproject.toml
[project]
name = "my-30daysofstreamlit-app"
version = "0.1.0"
requires-python = ">=3.11"
description = ""
dependencies = [
    "streamlit[snowflake]",
    "snowflake-ml-python"
]

[tool.uv]
# earlier versions of numba require a python version < 3.11
constraint-dependencies = ["numba>=0.56.0"]
